+++ libs/expo/betterangels/src/lib/screens/AddEditClient/AddEditClient.graphql
@@ -38,3 +38,19 @@ query GetClientProfile($id: ID!) {
     }
   }
 }
+
+mutation CreateClientProfile($data: CreateClientProfileInput!) {
+  createClientProfile(data: $data) {
+    ... on OperationInfo {
+      messages {
+        kind
+        field
+        message
+      }
+    }
+
+    ... on ClientProfileType {
+      id
+    }
+  }
+}
