+++ libs/expo/betterangels/src/lib/screens/index.ts
@@ -1,7 +1,7 @@
 export { default as AddClient } from './AddClient';
+export { default as AddEditClient } from './AddEditClient';
 export { default as Client } from './Client';
 export { default as Clients } from './Clients';
-export { default as EditClient } from './EditClient';
 export { default as Home } from './Home';
 export { default as Interactions } from './Interactions';
 export { default as Note } from './Note';
